metadata-ms/requirements.txt
fastapi
uvicorn[standard]
pydantic
cassandra-driver
httpx